+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Each change log are assumed to be accompanied by associated unit test and documentation updates.
 
+## master
+
+- refactored promise code to make it work with changes that were made to angular 1.3.0-beta.18
+- added .editorconfig file
+
 ## 0.5.0
 
 - initial release with change log file